Remove logical-pathnames.lisp
I had high hopes for logical pathnames, in terms of the debugger being able to
find the source code on a deployed version of CLPM to allow for easy
debugging. But the truth is, that doesn't work. Bugs are likely often also
going to include external libraries which don't have logical pathnames. If we
want to have debugging of deployed versions of CLPM, users need to compile it
themselves.