Simplify printing of contexts
It was a fun and enlightening experiment to use the pretty printer for
printing contexts, but it's ultimately unnecessary and just makes the code more
difficult to read.

If we want to go back to the pretty printer in the future, I think it's
ultimately better to rebind the pprint dispatch table and add new dispatches
for the types of things that are being printed.