Redirect clpm stderr to file when calling from client
The client did not continuously read from both CLPM's stderr and
stdout. Therefore, the stderr buffer could get filled (e.g., if logging
verbosity were increased or a system was being grovelled that generated a lot
of warnings) and clpm would have been unable to make forward progress.

Now, redirect stderr to a temporary file, reading it as necessary to report
errors in conditions or to the error dribble destination.